% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify-commata.R
\name{prettify2}
\alias{prettify2}
\title{Prettify tokenized output}
\usage{
prettify2(
  tbl,
  col = "feature",
  into = gibasa::get_dict_features("ipa"),
  col_select = seq_along(into)
)
}
\arguments{
\item{tbl}{A data.frame that has feature column to be prettified.}

\item{col}{<\code{\link[rlang:args_data_masking]{data-masked}}>
Column containing features to be prettified.}

\item{into}{Character vector that is used as column names of
features.}

\item{col_select}{Character or integer vector that will be kept
in prettified features.}
}
\value{
A data.frame.
}
\description{
Turns a single character column into features
while separating with delimiter.
}
